pyproject.toml
[tool.poetry]
name = "Akane"
version = "0.1.0"
description = ""
authors = ["Yoiduki <90896806+y-muen@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.1.0"
numpy = "^1.24.1"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
ipykernel = "^6.20.0"
pytest = "^7.2.0"
pycodestyle = "^2.10.0"
pydocstyle = "^6.2.2"
autopep8 = "^2.0.1"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"